# swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "UiRealtyPals API",
    "description": "Full REST API for UiRealtyPals - AI-powered real estate discovery platform for Noida. Auth: users send Authorization: Bearer <supabase-access-token>. Admin uses admin_session cookie. Webhooks use X-Webhook-Secret header.",
    "version": "1.0.0",
    "contact": {
      "name": "RealtyPals Engineering",
      "url": "https://realty-pals-dev-frontend-three.vercel.app"
    },
    "license": {
      "name": "Proprietary"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Local dev server"
    },
    {
      "url": "https://realtypalsdev.onrender.com",
      "description": "Production (Render)"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Supabase access token. Obtain via supabase.auth.signIn*(). Send as: Authorization: Bearer <token>"
      },
      "AdminSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "admin_session",
        "description": "Admin session cookie set by POST /api/v1/admin/login. Required for all /admin/* routes."
      },
      "WebhookSecret": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Webhook-Secret",
        "description": "HMAC-SHA256 secret for webhook lead ingestion."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "example": "Not found"
          }
        }
      },
      "ValidationError": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          },
          "details": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "path": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "UnitType": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "project_id": {
            "type": "string",
            "format": "uuid"
          },
          "bhk": {
            "type": "integer",
            "example": 3
          },
          "price_min_cr": {
            "type": "number",
            "example": 1.85,
            "description": "Minimum price in Crore INR"
          },
          "price_max_cr": {
            "type": "number",
            "example": 3.14
          },
          "super_area_sqft": {
            "type": "integer",
            "example": 1650
          },
          "carpet_area_sqft": {
            "type": "integer",
            "example": 1250
          }
        }
      },
      "ProjectImage": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "url": {
            "type": "string",
            "format": "uri"
          },
          "type": {
            "type": "string",
            "enum": [
              "hero",
              "gallery",
              "floor_plan",
              "amenity"
            ]
          },
          "sort_order": {
            "type": "integer"
          }
        }
      },
...
      "ProjectCard": {
        "type": "object",
        "description": "Lightweight project card for listing and chat responses",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "example": "Godrej Meridien"
          },
          "slug": {
            "type": "string",
            "example": "godrej-meridien-sector-106"
          },
          "tagline": {
            "type": "string"
          },
          "sector": {
            "type": "string",
            "example": "Sector 106"
          },
          "city": {
            "type": "string",
            "example": "Noida"
          },
          "status": {
            "type": "string",
            "enum": [
              "under_construction",
              "ready_to_move",
              "new_launch"
            ]
          },
          "possession_date": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "rera_number": {
            "type": "string",
            "nullable": true
          },
          "total_units": {
            "type": "integer",
            "nullable": true
          },
          "total_towers": {
            "type": "integer",
            "nullable": true
          },
          "land_area_acres": {
            "type": "number",
            "nullable": true
          },
          "builder": {
            "type": "object",
            "properties": {
              "name": {
                "type": "string"
              },
              "slug": {
                "type": "string"
              }
            }
          },
          "unit_types": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UnitType"
            }
          },
          "images": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProjectImage"
            }
          }
        }
      },
...
      "Document": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "storage_url": {
            "type": "string",
            "format": "uri"
          },
          "doc_type": {
            "type": "string",
            "enum": [
              "rera",
              "brochure",
              "floor_plan",
              "price_list",
              "payment_plan",
              "legal",
              "other"
            ]
          },
          "file_size_bytes": {
            "type": "integer",
            "nullable": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    },
    "parameters": {
      "SlugPath": {
        "name": "slug",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        },
        "description": "URL-safe slug identifier"
      },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "Unauthorized - valid Bearer token or admin session required",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "BadRequest": {
        "description": "Invalid request body or query parameters",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/v1/projects": {
      "get": {
        "tags": [
          "Projects"
        ],
        "summary": "List projects",
        "description": "Filtered project list (max 20). Cached Redis 5 min. All params optional.",
        "operationId": "listProjects",
        "parameters": [
          {
            "name": "sector",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "example": "Sector 150",
            "description": "Case-insensitive partial match"
          },
          {
            "name": "city",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "Noida"
            }
          },
          {
            "name": "bhk",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "example": 3,
            "description": "BHK count filter"
          },
          {
            "name": "budget_max_cr",
            "in": "query",
            "schema": {
              "type": "number"
            },
            "example": 2.5,
            "description": "Max price in Crore INR"
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "under_construction",
                "ready_to_move",
                "new_launch"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Project list",
            "headers": {
              "X-Cache": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "HIT",
                    "MISS"
                  ]
                },
                "description": "Redis cache status"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "projects": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ProjectCard"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
    "/api/v1/projects/{slug}": {
      "get": {
        "tags": [
          "Projects"
        ],
        "summary": "Get project detail",
        "description": "Full detail: DNA, decision profile, recommendation profile, persona, competitors, amenities, connectivity, computed recommendation_score. Cached 15 min. Internal DNA raw scores stripped from response.",
        "operationId": "getProjectBySlug",
        "parameters": [
          {
            "$ref": "#/components/parameters/SlugPath"
          }
        ],
        "responses": {
          "200": {
            "description": "Project detail",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "project": {
                      "$ref": "#/components/schemas/ProjectDetail"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/api/v1/projects/{slug}/documents": {
      "get": {
        "tags": [
          "Projects",
          "Documents"
        ],
        "summary": "List project documents",
        "description": "RERA, brochures, floor plans. Ordered newest first.",
        "operationId": "getProjectDocuments",
        "parameters": [
          {
            "$ref": "#/components/parameters/SlugPath"
          }
        ],
        "responses": {
          "200": {
            "description": "Documents",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "documents": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Document"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/api/v1/projects/{slug}/payment-plan": {
      "get": {
        "tags": [
          "Projects"
        ],
        "summary": "Get payment plan",
        "description": "Verified payment schedule or available:false with guidance.",
        "operationId": "getProjectPaymentPlan",
        "parameters": [
          {
            "$ref": "#/components/parameters/SlugPath"
          }
        ],
        "responses": {
          "200": {
            "description": "Payment plan or unavailability notice",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "available": {
                      "type": "boolean"
                    },
                    "plan": {
                      "type": "object",
                      "nullable": true
                    },
                    "message": {
                      "type": "string",
                      "nullable": true
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/api/v1/projects/{slug}/cost-sheet": {
      "get": {
        "tags": [
          "Projects"
        ],
        "summary": "Get cost sheet",
        "description": "Verified cost breakdown (GST, stamp duty, registration, IFMS, club, parking) + illustrative total for representative unit.",
        "operationId": "getProjectCostSheet",
        "parameters": [
          {
            "$ref": "#/components/parameters/SlugPath"
          }
        ],
        "responses": {
          "200": {
            "description": "Cost sheet",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "available": {
                      "type": "boolean"
                    },
                    "sheet": {
                      "type": "object",
                      "nullable": true
                    },
                    "illustration": {
                      "nullable": true,
                      "type": "object",
                      "properties": {
                        "base_price_cr": {
                          "type": "number"
                        },
                        "gst_cr": {
                          "type": "number"
                        },
                        "stamp_duty_cr": {
                          "type": "number"
                        },
                        "registration_cr": {
                          "type": "number"
                        },
                        "parking_cr": {
                          "type": "number"
                        },
                        "ifms_cr": {
                          "type": "number"
                        },
                        "club_cr": {
                          "type": "number"
                        },
                        "total_cost_cr": {
                          "type": "number"
                        }
                      }
                    },
                    "illustration_note": {
                      "type": "string",
                      "nullable": true
                    },
                    "message": {
                      "type": "string",
                      "nullable": true
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/api/v1/projects/{slug}/investment": {
      "get": {
        "tags": [
          "Projects"
        ],
        "summary": "Get investment intelligence",
        "description": "AI-derived investment indicators from DNA scores - never LLM-fabricated. Always includes disclaimer.",
        "operationId": "getProjectInvestment",
        "parameters": [
          {
            "$ref": "#/components/parameters/SlugPath"
          }
        ],
        "responses": {
          "200": {
            "description": "Investment intelligence",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "available": {
                      "type": "boolean"
                    },
                    "intelligence": {
                      "nullable": true,
                      "type": "object",
                      "properties": {
                        "sector": {
                          "type": "string"
                        },
                        "status": {
                          "type": "string"
                        },
                        "possession_date": {
                          "type": "string",
                          "format": "date",
                          "nullable": true
                        },
                        "investment_thesis": {
                          "type": "string",
                          "nullable": true
                        },
                        "investor_thesis": {
                          "type": "string",
                          "nullable": true
                        },
                        "potential_appreciation": {
                          "type": "string",
                          "enum": [
                            "Strong",
                            "Moderate",
                            "Weak"
                          ],
                          "nullable": true
                        },
                        "data_note": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
    "/api/v1/admin/projects": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Admin: list all projects",
        "operationId": "adminListProjects",
        "security": [
          {
            "AdminSession": []
          }
        ],
        "responses": {
          "200": {
            "description": "Full project list"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Admin: create project",
        "operationId": "adminCreateProject",
        "security": [
          {
            "AdminSession": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "description": "Project creation payload (mirrors Prisma Project model)"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/v1/admin/projects/{id}": {
      "put": {
        "tags": [
          "Admin"
        ],
        "summary": "Admin: update project",
        "operationId": "adminUpdateProject",
        "security": [
          {
            "AdminSession": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "delete": {
        "tags": [
          "Admin"
        ],
        "summary": "Admin: delete project",
        "operationId": "adminDeleteProject",
        "security": [
          {
            "AdminSession": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "get": {
        "summary": "Auto-generated endpoint for GET /api/v1/admin/projects/:id",
        "tags": [
          "Admin"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          }
        },
        "security": [
          {
            "AdminSession": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      },
      "patch": {
        "summary": "Auto-generated endpoint for PATCH /api/v1/admin/projects/:id",
        "tags": [
          "Admin"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          }
        },
        "security": [
          {
            "AdminSession": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/projects/{id}/cost-sheet": {
      "put": {
        "summary": "Auto-generated endpoint for PUT /api/v1/admin/projects/:id/cost-sheet",
        "tags": [
          "Admin"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          }
        },
        "security": [
          {
            "AdminSession": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/projects/{id}/documents": {
      "get": {
        "summary": "Auto-generated endpoint for GET /api/v1/admin/projects/:id/documents",
        "tags": [
          "Admin"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          }
        },
        "security": [
          {
            "AdminSession": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/api/v1/admin/projects/{id}/payment-plan": {
      "put": {
        "summary": "Auto-generated endpoint for PUT /api/v1/admin/projects/:id/payment-plan",
        "tags": [
          "Admin"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          }
        },
        "security": [
          {
            "AdminSession": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        }
      }
    },
...
  }
}